Commit (Work in progress)
Using the images now so I'm going to create this commit so I can use the new system GT uses
