+++ Slippery Ice/PluginInfo.cs
@@ -1,0 +1,9 @@
+namespace SlipperyIce
+{
+    class PluginInfo
+    {
+        public const string GUID = "com.fchb1239.gorillatag.slipperyice";
+        public const string Name = "Slippery Ice";
+        public const string Version = "1.0.3";
+    }
+}
